+++ craft_parts/executor/part_handler.py
@@ -19,7 +19,7 @@
 import logging
 import os
 import shutil
-from collections.abc import Callable, Iterator, Mapping, Sequence
+from collections.abc import Callable, Iterable, Iterator, Mapping, Sequence
 from contextlib import contextmanager
 from pathlib import Path
 from typing import Any, cast
@@ -49,7 +49,7 @@ from craft_parts.state_manager import (
 from craft_parts.state_manager.stage_state import StageState
 from craft_parts.steps import Step
 from craft_parts.utils import file_utils, os_utils
-from craft_parts.utils.partition_utils import DEFAULT_PARTITION
+from craft_parts.utils.partition_utils import BUILD_PARTITION, DEFAULT_PARTITION
 
 from . import filesets, migration
 from .environment import generate_step_environment
@@ -229,6 +229,7 @@ class PartHandler:
         overlay_manager: OverlayManager,
         ignore_patterns: list[str] | None = None,
         base_layer_hash: LayerHash | None = None,
+        build_environment: Iterable[str] | None = None,
     ) -> None:
         self._part = part
         self._part_info = part_info
@@ -236,7 +237,7 @@ class PartHandler:
         self._track_stage_packages = track_stage_packages
         self._overlay_manager = overlay_manager
         self._base_layer_hash = base_layer_hash
-        self._app_environment: dict[str, str] = {}
+        self._build_environment = build_environment
 
         self._plugin = plugins.get_plugin(
             part=part,
@@ -498,10 +499,15 @@ class PartHandler:
             # time around. We can be confident that this won't overwrite anything else,
             # because to do so would require changing the `organize` keyword, which will
             # make the build step dirty and require a clean instead of an update.
+            organize_install_dirs = {
+                **self._part.part_install_dirs,
+                BUILD_PARTITION: self._part.part_build_dir,
+            }
+
             organize_files(
                 part_name=self._part.name,
                 file_map=self._part.spec.organize_files,
-                install_dir_map=self._part.part_install_dirs,
+                install_dir_map=organize_install_dirs,
                 overwrite=update,
                 default_partition=step_info.default_partition,
             )
@@ -540,6 +546,8 @@ class PartHandler:
         """
         self._make_dirs()
 
+        self._migrate_overlay_files_to_stage()
+
         contents = self._run_step(
             step_info=step_info,
             scriptlet_name="override-stage",
@@ -547,8 +555,6 @@ class PartHandler:
             stdout=stdout,
             stderr=stderr,
         )
-
-        self._migrate_overlay_files_to_stage()
 
         # Overlay integrity is checked based by the hash of its last (topmost) layer,
         # so we compute it for all parts. The overlay hash is added to the stage state
@@ -597,6 +603,8 @@ class PartHandler:
         """
         self._make_dirs()
 
+        self._migrate_overlay_files_to_prime()
+
         contents = self._run_step(
             step_info=step_info,
             scriptlet_name="override-prime",
@@ -605,7 +613,6 @@ class PartHandler:
             stderr=stderr,
         )
 
-        self._migrate_overlay_files_to_prime()
         default_partition = self._part_info.default_partition or DEFAULT_PARTITION
 
         primed_stage_packages: set[str]
@@ -661,6 +668,9 @@ class PartHandler:
         )
 
         if step_info.step == Step.BUILD:
+            # Prepend environment set by the application.
+            step_env = self._prepend_build_environment(step_env)
+
             # Validate build environment. Unlike the pre-validation we did in
             # the execution prologue, we don't assume that a different part
             # can add elements to the build environment. All part dependencies
@@ -1363,6 +1373,24 @@ class PartHandler:
 
         for snap_source in snap_sources:
             snap_source.provision(install_dir, keep=True)
+
+    def _prepend_build_environment(self, content: str) -> str:
+        if not self._build_environment:
+            return content
+
+        current_dir = Path.cwd()
+        try:
+            # Set the current working directory to the build directory. Build
+            # environment generators may produce different results depending
+            # on the current path.
+            os.chdir(self._part.part_build_dir)
+            env_list = ["# Build environment from application"]
+            env_list.extend([f"export {x}" for x in self._build_environment])
+            env_list.extend(["", content])
+        finally:
+            os.chdir(current_dir)
+
+        return "\n".join(env_list)
 
 
 def _remove(filename: Path) -> None: