+++ craft_parts/lifecycle_manager.py
@@ -19,7 +19,7 @@
 import os
 import re
 import sys
-from collections.abc import Sequence
+from collections.abc import Iterable, Sequence
 from pathlib import Path
 from typing import Any, cast
 
@@ -89,7 +89,9 @@ class LifecycleManager:
     :param usrmerged_by_default: Whether the parts' install dirs should be filled with
         usrmerge-safe directories and symlinks prior to a part's build.
     :param use_host_sources: Whether overlay steps should also include the repository
-      sources defined on the host.
+        sources defined on the host.
+    :param build_environment: An iterable of environment variables in name=value format
+        to be set during the build step.
     :param custom_args: Any additional arguments that will be passed directly
         to callbacks.
     """
@@ -120,6 +122,7 @@ class LifecycleManager:
         filesystem_mounts: dict[str, Any] | None = None,
         usrmerged_by_default: bool = False,
         use_host_sources: bool = False,
+        build_environment: Iterable[str] | None = None,
         **custom_args: Any,  # custom passthrough args
     ) -> None:
         # pylint: disable=too-many-locals
@@ -226,6 +229,7 @@ class LifecycleManager:
             base_layer_dir=base_layer_dir,
             base_layer_hash=layer_hash,
             use_host_sources=use_host_sources,
+            build_environment=build_environment,
         )
         self._project_info = project_info
         # pylint: enable=too-many-locals