+++ tests/unit/executor/test_part_handler.py
@@ -15,6 +15,8 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 import logging
+import textwrap
+from collections.abc import Generator, Iterable
 from pathlib import Path
 from typing import Any, cast
 from unittest.mock import call
@@ -136,6 +138,27 @@ class TestPartHandling:
 
         assert self._mock_mount_overlayfs.mock_calls == []
         assert self._mock_umount.mock_calls == []
+
+    def test_run_build_passes_build_dir_to_organize(self, mocker):
+        mocker.patch("craft_parts.executor.step_handler.StepHandler._builtin_build")
+        mocker.patch(
+            "craft_parts.packages.Repository.get_installed_packages",
+            return_value=[],
+        )
+        mocker.patch(
+            "craft_parts.packages.snaps.get_installed_snaps",
+            return_value=[],
+        )
+        mocker.patch("subprocess.check_output", return_value=b"os-info")
+        mock_organize = mocker.patch("craft_parts.executor.part_handler.organize_files")
+
+        self._handler._run_build(
+            StepInfo(self._part_info, Step.BUILD), stdout=None, stderr=None
+        )
+
+        organize_install_dirs = mock_organize.call_args.kwargs["install_dir_map"]
+        assert organize_install_dirs["build"] == self._part.part_build_dir
+        assert "build" not in self._part.part_install_dirs
 
     @pytest.mark.usefixtures("new_dir")
     @pytest.mark.parametrize("out_of_source", [True, False])
@@ -1323,6 +1346,7 @@ class TestDirs:
         new_dir,
         plugin: str = "make",
         part_spec: dict[str, Any] | None = None,
+        build_environment: Iterable[str] | None = None,
     ) -> tuple[Part, PartHandler]:
         spec = part_spec or {}
         part = Part(
@@ -1345,6 +1369,7 @@ class TestDirs:
             part_info=part_info,
             part_list=[part],
             overlay_manager=ovmgr,
+            build_environment=build_environment,
         )
 
         return part, handler
@@ -1416,3 +1441,47 @@ class TestDirs:
         handler._make_dirs()
 
         assert list(part.part_install_dir.iterdir()) == []
+
+    def test_build_environment(self, new_dir, partitions):
+        """Test application-specified build environment variables."""
+        part, handler = self._part_and_handler(
+            usrmerged_by_default=True,
+            new_dir=new_dir,
+            partitions=partitions,
+            plugin="nil",
+            build_environment=['FOO="foo value"', 'BAR="bar value"'],
+        )
+
+        handler._make_dirs()
+        step_env = handler._prepend_build_environment("content")
+
+        assert step_env == textwrap.dedent("""\
+            # Build environment from application
+            export FOO="foo value"
+            export BAR="bar value"
+
+            content""")
+
+    def test_build_environment_from_generator(self, new_dir, partitions):
+        """Test application-specified build environment variables."""
+
+        def envgen() -> Generator[str]:
+            yield from ['FOO="foo value"', 'BAR="bar value"']
+
+        part, handler = self._part_and_handler(
+            usrmerged_by_default=True,
+            new_dir=new_dir,
+            partitions=partitions,
+            plugin="nil",
+            build_environment=envgen(),
+        )
+
+        handler._make_dirs()
+        step_env = handler._prepend_build_environment("content")
+
+        assert step_env == textwrap.dedent("""\
+            # Build environment from application
+            export FOO="foo value"
+            export BAR="bar value"
+
+            content""")