+++ docs/reference/changelog.rst
@@ -19,23 +19,73 @@ Changelog
 
   For a complete list of commits, check out the `X.Y.Z`_ release on GitHub.
 
-.. _release-2.34.0:
+Unreleased
+----------
 
-2.34.0 (Unreleased)
+Bug fixes:
+
+- The Overlay and Build steps of parts that organize content to the overlay
+  are executed before these steps in other parts, even when ``after`` is
+  used.
+
+.. _release-2.33.1:
+
+2.33.1 (2026-06-17)
+-------------------
+
+Bug fixes:
+
+- Always migrate overlay files to Stage and Prime first, even if the part being staged
+  or primed also has contents coming from the install directory.
+
+For a complete list of commits, check out the `2.33.1`_ release on GitHub.
+
+.. _release-2.35.0:
+
+2.35.0 (2026-06-17)
 -------------------
 
 New features:
 
+- Applications can define default parameters for autoconf and make when subclassing
+  the plugins.
+- The plus sign is allowed in partition names. More restrictive rules must be
+  enforced at application level.
+
+Bug fixes:
+
+- Always migrate overlay files to Stage and Prime first, even if the part being staged
+  or primed also has contents coming from the install directory.
+
+For a complete list of commits, check out the `2.35.0`_ release on GitHub.
+
+.. _release-2.34.0:
+
+2.34.0 (2026-06-10)
+-------------------
+
+New features:
+
+- Plugins can now define overlay commands.
+- Add support for organizing files from the build directory using the ``(build)``
+  pseudo-partition.
 - Add a ``gradle-use`` plugin for publishing Gradle artifacts to a local Maven
   repository.
 - For explicit typing of file system operations, switch to the `pathlib module
   <https://docs.python.org/3/library/pathlib.html>`_. All uses of ``str`` for paths are
   replaced with ``pathlib.Path``. In the public APIs, this change only impacts the
   ``files`` and ``directories`` variables of the ``MigrationState`` class.
+- Allow applications to set up build environment variables when creating the
+  lifecycle manager.
 
 Bug fixes:
 
+- Reject organize source entries that resolve outside the part install directory.
+- Fix edge cases when organizing files to the overlay partition.
+- Track ``override-overlay`` changes in part state.
 - Wrap streaming request errors for file sources in ``NetworkRequestError``.
+
+For a complete list of commits, check out the `2.34.0`_ release on GitHub.
 
 .. _release-2.33.0:
 
@@ -1712,6 +1762,9 @@ For a complete list of commits, check out the `2.0.0`_ release on GitHub.
 .. _craft-cli issue #172: https://github.com/canonical/craft-cli/issues/172
 .. _Poetry: https://python-poetry.org
 
+.. _2.35.0: https://github.com/canonical/craft-parts/releases/tag/2.35.0
+.. _2.34.0: https://github.com/canonical/craft-parts/releases/tag/2.34.0
+.. _2.33.1: https://github.com/canonical/craft-parts/releases/tag/2.33.1
 .. _2.33.0: https://github.com/canonical/craft-parts/releases/tag/2.33.0
 .. _2.32.0: https://github.com/canonical/craft-parts/releases/tag/2.32.0
 .. _2.31.0: https://github.com/canonical/craft-parts/releases/tag/2.31.0